Globalised element formatting across pages, reworked news page
Now the majority of the styling is all in `style.css`, as it should be.

...However, nearly everything uses `rem` measurements, which is only a step
above using pixels. The issue is that it is way too late - both as in
"within this project" and "within today" - for me to really care anymore.
It "works", even though the news has to have surperfluous font-size
redefinitions, when it could be using a single one that would be propagated
through the relative `em` sizings. Whatever.

Secondly, the news page now uses a nice column system instead of a grid,
letting articles be whatever height they want without disturbing each other.
I also moved the publish dates to be next to the author, since it feels more
appropriate to me - ...and because it means there's one less text style to
worry about "- _-)

To be fair, I've enjoyed most of this experience, but GOD I really can't wait
for this to all be over soon.